RC2 added controller tests for empty fields values and delete method

+++ msky_money_persistence/src/main/java/com/dsu/dao/api/AbstractCrudDao.java
@@ -108,6 +108,8 @@ public abstract class AbstractCrudDao<I> implements CrudDao<I> {
 
 	public void delete(Long id) {
 		I instance = findById(id);
-		entityManager.remove(instance);
+		if (instance != null) {
+			entityManager.remove(instance);
+		}
 	}
 }
